app/requirements.txt
aiofiles==22.1.0
aiohttp==3.8.3
aiosignal==1.3.1
altair==4.2.2
anyio==3.6.2
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
colorlover==0.3.0
comm==0.1.2
contourpy==1.0.7
cufflinks==0.17.3
cycler==0.11.0
datasets==2.8.0
dbus-python==1.2.16
debugpy==1.6.6
decorator==5.1.1
dill==0.3.6
distro==1.4.0
entrypoints==0.4
executing==1.2.0
fastapi==0.89.1
ffmpy==0.3.0
filelock==3.9.0
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
gradio==3.16.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.12.0
idna==3.4
importlib-resources==5.10.2
ipykernel==6.20.2
ipython==8.8.0
ipywidgets==8.0.4
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.17.3
jupyter-client==7.4.9
jupyter-core==5.1.5
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
linkify-it-py==1.0.3
markdown-it-py==2.1.0
MarkupSafe==2.1.2
matplotlib==3.6.3
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
multidict==6.0.4
multiprocess==0.70.14
nest-asyncio==1.5.6
numpy==1.24.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
orjson==3.8.5
packaging==23.0
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pkgutil-resolve-name==1.3.10
platformdirs==2.6.2
plotly==5.13.0
prompt-toolkit==3.0.36
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==10.0.1
pycryptodome==3.17
pydantic==1.10.4
pydub==0.25.1
Pygments==2.14.0
PyGObject==3.36.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2022.7.1
PyYAML==6.0
pyzmq==25.0.0
regex==2022.10.31
requests==2.28.2
responses==0.18.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
ssh-import-id==5.10
stack-data==0.6.2
starlette==0.22.0
tenacity==8.1.0
timm==0.6.12
tokenizers==0.13.2
toolz==0.12.0
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tornado==6.2
tqdm==4.64.1
traitlets==5.8.1
transformers==4.26.0
typing-extensions==4.4.0
uc-micro-py==1.0.1
urllib3==1.26.14
uvicorn==0.20.0
wcwidth==0.2.6
websockets==10.4
widgetsnbextension==4.0.5
xxhash==3.2.0
yarl==1.8.2
zipp==3.12.0
